Feito correções de conflito no projeto
