+++ app/backend/global_functions.py
@@ -16,6 +16,7 @@ def save_log(log_name: str) -> None:
     log_line = f'log: {log_name} - Data e hora: {current_formated_datetime}'
     save_in_database(log_line, 'logs/historico_log.txt')
 
+
 def read_marketplaces() -> list:
     marketplaces: list = []
     archive = open('logs/marketplaces.txt', 'r')
@@ -28,6 +29,10 @@ def read_marketplaces() -> list:
     
     archive.close()
     return marketplaces
+
+
+read_marketplaces() 
+
 
 def read_products() -> list:
     products: list = []
@@ -42,15 +47,36 @@ def read_products() -> list:
     archive.close()
     return products
 
+
+def read_categories() -> list:
+    categories: list = []
+    archive = open('../logs/categories.txt', 'r')
+
+    for obj in archive:
+        line_cleaned = obj.strip()  # clear caracteres and clear white spaces (\n \t \r ' ')
+        data_line = line_cleaned.split('"')
+        formated_line = {data_line[1]: data_line[3], data_line[5]: data_line[7]}
+        categories.append(formated_line)
+
+    archive.close()
+    return categories
+
+
 menu = [
     {'name': 'Marketplaces',
      'route': '/marketplaces'},
     {'name': 'Produtos',
      'route': '/products'},
-    {'name': 'Listar Marketplaces',
+    {'name': 'Categorias',
+     'route': '/categories'},
+     {'name': 'Listar Marketplace',
      'route': '/list-marketplaces'},
     {'name': 'Listar Produtos',
      'route': '/list-products'},
+    {'name': 'Listar Categorias',
+     'route': '/list-categories'},
+    {'name': 'Log de uso',
+     'route': '/historico'}
 ]
 
 links = [
