+++ app/frontend/templates/categories.html
