+++ app/frontend/templates/list_categories.html
