+++ app/logs/historico_log.txt
@@ -60,3 +60,4 @@ log: Salvo marketplace - Mercado Livre - Data e hora: 06/01/2021 16:54:01
 log: Salvo produto - PS5 - Data e hora: 06/01/2021 16:54:40
 log: Salvo marketplace - viavarejo - Data e hora: 06/01/2021 17:31:59
 log: Salvo produto - teclado - Data e hora: 06/01/2021 17:32:14
+log: Salvo marketplace - LUCAS TESTE - Data e hora: 07/01/2021 11:08:27
