--- logs/categories.txt
@@ -1,2 +1,0 @@
-"name": "Lucas Teste categoria", "description": "teste"
-"name": "Lucas Teste categoria2", "description": "ola mundo"
