--- logs/historico_log.txt
@@ -1,41 +1,0 @@
-log: Salvo marketplace - None - Data e hora: 05/01/2021 09:47:31
-log: Salvo marketplace - 123 - Data e hora: 05/01/2021 09:47:36
-log: Salvo produto - None - Data e hora: 05/01/2021 09:47:39
-log: Salvo produto - 456 - Data e hora: 05/01/2021 09:47:46
-log: None - Data e hora: 05/01/2021 10:01:25
-log: 456789 - Data e hora: 05/01/2021 10:01:30
-log: Salvo marketplace - None - Data e hora: 05/01/2021 10:03:12
-log: Salvo marketplace - aaa - Data e hora: 05/01/2021 10:03:16
-log: Salvo marketplace - None - Data e hora: 05/01/2021 10:05:22
-log: Salvo marketplace - zzz - Data e hora: 05/01/2021 10:05:25
-log: Salvo produto - None - Data e hora: 05/01/2021 10:05:27
-log: Salvo produto - xxx - Data e hora: 05/01/2021 10:05:33
-log: Salvo produto - None - Data e hora: 05/01/2021 10:09:07
-log: Salvo produto - zxczxc - Data e hora: 05/01/2021 10:09:11
-log: Salvo produto - None - Data e hora: 05/01/2021 10:13:02
-log: Salvo produto - 456789 - Data e hora: 05/01/2021 10:13:08
-log: Salvo marketplace - None - Data e hora: 05/01/2021 10:13:35
-log: Salvo marketplace - asdas - Data e hora: 05/01/2021 10:13:38
-log: Salvo produto - None - Data e hora: 05/01/2021 10:13:40
-log: Salvo produto - asdfasdfasd - Data e hora: 05/01/2021 10:13:45
-log: Salvo marketplace - None - Data e hora: 05/01/2021 10:20:26
-log: Salvo marketplace - teste - Data e hora: 05/01/2021 10:20:31
-log: Salvo produto - None - Data e hora: 05/01/2021 10:20:32
-log: Salvo produto - teste prod - Data e hora: 05/01/2021 10:20:40
-log: Salvo marketplace - None - Data e hora: 05/01/2021 10:22:19
-log: Salvo marketplace - None - Data e hora: 05/01/2021 17:53:06
-log: Salvo produto - None - Data e hora: 05/01/2021 17:53:12
-log: Salvo produto - None - Data e hora: 05/01/2021 17:53:16
-log: Salvo produto - None - Data e hora: 05/01/2021 17:53:30
-log: Salvo produto - None - Data e hora: 06/01/2021 10:45:39
-log: Salvo produto - lucas - Data e hora: 06/01/2021 10:45:46
-log: Salvo produto - None - Data e hora: 06/01/2021 11:07:36
-log: Salvo produto - lucas - Data e hora: 06/01/2021 11:07:44
-log: Salvo produto - None - Data e hora: 06/01/2021 11:16:15
-log: Salvo category - None - Data e hora: 06/01/2021 19:46:54
-log: Salvo marketplace - None - Data e hora: 06/01/2021 19:46:59
-log: Salvo category - None - Data e hora: 06/01/2021 19:47:01
-log: Salvo category - Lucas Teste categoria - Data e hora: 06/01/2021 19:47:18
-log: Salvo category - Lucas Teste categoria - Data e hora: 06/01/2021 19:48:57
-log: Salvo category - Lucas Teste categoria2 - Data e hora: 06/01/2021 19:49:16
-log: Salvo produto - None - Data e hora: 06/01/2021 19:55:42
