--- logs/marketplaces.txt
@@ -1,17 +1,0 @@
-'name': None, 'description': None
-'name': 123, 'description': 123
-'name': None, 'description': None
-'name': 456789, 'description': 45789
-'name': None 'description': None
-'name': aaa 'description': a
-'name': None 'description': None
-'name': zzz 'description': zzz
-'name': None, 'description': None
-'name': zxczxc, 'description': zxczxc
-'name': None, 'description': None
-'name': asdas, 'description': asdasd
-'name': None, 'description': None
-'name': teste, 'description': teste
-'name': None, 'description': None
-'name': None, 'description': None
-'name': None, 'description': None
